Refactor: extract CookieNotice as component
